+++ src/main/java/functionLayer/Product.java
@@ -17,12 +17,12 @@ public class Product {
     private double price;
     private double qty;
     private double priceLine;
-    private double length;
-    private double width;
-    private double height;
+    private int length;
+    private int width;
+    private int height;
     private double totalPriceOfOrder;
 
-    public Product(int id, String produktName, String category, double price, double length, double width, double height) {
+    public Product(int id, String produktName, String category, double price, int length, int width, int height) {
         this.id = id;
         this.produktName = produktName;
         this.category = category;
@@ -32,7 +32,7 @@ public class Product {
         this.height = height;
     }
 
-    public Product(int id, String produktName, String category, double price, double qty, double priceLine, double length, double width, double height) {
+    public Product(int id, String produktName, String category, double price, double qty, double priceLine, int length, int width, int height) {
         this.id = id;
         this.produktName = produktName;
         this.category = category;
@@ -82,15 +82,15 @@ public class Product {
         return qty;
     }
 
-    public double getLength() {
+    public int getLength() {
         return length;
     }
 
-    public double getWidth() {
+    public int getWidth() {
         return width;
     }
 
-    public double getHeight() {
+    public int getHeight() {
         return height;
     }
